New libraries connected and pages are laid out
