+++ .stylelintrc.yaml
@@ -1,0 +1,4 @@
+extends: 
+  [
+    "stylelint-prettier/recommended",
+  ]
