+++ JS/game.js
@@ -1,0 +1,2 @@
+import { contentGame } from './contentGame.js';
+contentGame();
