+++ JS/index.js
@@ -1,0 +1,3 @@
+import { buttonStart, levelDifficulty } from './difficultySelection.js';
+levelDifficulty();
+buttonStart();
