+++ game.html
@@ -4,9 +4,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <script type="module" src="game.js"></script>
+    <link rel="stylesheet" href="style.css">
 </head>
-<body>
-    <div class="app"></div>
+<body class="game-body">
+    <script type="module" src="JS/game.js"></script>
 </body>
 </html>
