--- game.js
@@ -1,2 +1,0 @@
-import { contentGame } from "./contentGame.js";
-contentGame();
