+++ index.html
@@ -5,9 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
     <link rel="stylesheet" href="style.css">
-    <script type="module" src="index.js"></script>
 </head>
-<body>
+<body class="body">
         <div class="general">
             <div class="general_level">
                 <h1 class="general_level_h1">Выбери сложность</h1>
@@ -21,5 +20,6 @@
                 <button class="general_button">Старт</button>
             </div>
         </div>
+        <script type="module" src="JS/index.js"></script>
 </body>
 </html>
