--- index.js
@@ -1,5 +1,0 @@
-import { buttonStart, levelDifficulty } from "./difficultySelection.js";
-levelDifficulty();
-buttonStart();
-
-
